Implement AbstractRamLogger
